finexio-matcher/pyproject.toml
[project]
name = "finexio-payee-matcher"
version = "1.0.0"
description = "Production-ready payee matching service with hybrid deterministic + ML approach"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=23.2.0",
    "rapidfuzz>=3.5.0",
    "metaphone>=0.6",
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "prometheus-client>=0.19.0",
    "openai>=1.0.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "google-cloud-bigquery>=3.13.0",
    "db-dtypes>=1.2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]